Review: clean up trap handler after implementation
- trap.c: panic on unexpected interrupts instead of silently
  returning (prevents infinite re-trap loops for unhandled
  interrupt sources); use if/else for interrupt vs exception
- main.c: explicit (uint64) cast on kernelvec for csrw clarity
- .clang-format: fix run/check command examples

+++ kernel/main.c
@@ -24,7 +24,7 @@ kmain(void) {
 
     /* Set stvec to point to kernelvec (direct mode).
      * This must happen before anything that could trap. */
-    csrw(stvec, kernelvec);
+    csrw(stvec, (uint64)kernelvec);
 
     kprintf("\n");
     kprintf("bobchouOS is booting...\n");
